# livros-next/app/model/Editora.ts
class Editora {
    codigo: string;
    codEditora: number;
    nome: string;

    constructor(codigo: string, codEditora: number, nome: string) {
        this.codigo = codigo;
        this.codEditora = codEditora;
        this.nome = nome;
    }
}

export default Editora;