# livros-react/src/model/Livro.js
class Livro {
    constructor(codigo, codEditora, titulo, resumo, autores) {
        this.codigo = codigo;
        this.codEditora = codEditora;
        this.titulo = titulo;
        this.resumo = resumo;
        this.autores = autores;
    }
}

export default Livro;